Allow hardcoding Torrent.infohash
When creating a Torrent instance from a magnet URI, we cannot calculate the info
hash because the metainfo is lacking most of the needed information.

This is a bit dirty, but it's natural to assume that the Torrent knows the info
hash if the magnet URI from which it is created knows it.