Reader._fake_read_file(): Maybe report first and last piece
Instead of sending `None`, the first and last piece of the faked file might have
to be actual bytes so the hash can mismatch and an error is reported for those
pieces.

This is not the case, if the last piece of the previous file or the last piece
of the faked file stops at the piece boundary.